Fix data type bug causing unit power factor to lose precision.
See #40746

Reported by: kevin in Korea <kevin.in.korea@gmail.com>

[[originally from svn r15577]]

+++ common/unit.c
@@ -429,7 +429,7 @@ bool can_unit_change_homecity(const struct unit *punit)
 **************************************************************************/
 int get_activity_rate(const struct unit *punit)
 {
-  int fact = unit_type(punit)->veteran[punit->veteran].power_fact;
+  double fact = unit_type(punit)->veteran[punit->veteran].power_fact;
 
   /* The speed of the settler depends on its base move_rate, not on
    * the number of moves actually remaining or the adjusted move rate.
